removed unnecessary memoization
