+++ package.json
@@ -39,7 +39,6 @@
     "html-webpack-plugin": "2.29.0",
     "jsonwebtoken": "^8.2.1",
     "jwt-decode": "^2.2.0",
-    "lodash.memoize": "4.1.2",
     "mongoose": "^5.1.0",
     "object-assign": "4.1.1",
     "passport": "^0.4.0",
